detect: add tcp.session keyword for unified TCP lifecycle matching
Introduce the tcp.session: keyword (Redmine #7704, parameter portion).
The keyword accepts a comma-separated subset of {setup, established,
closing} and matches packets whose TCP session is in any of the named
lifecycle phases.

Implementation:
- Parser with structural comma validation (rejects leading, trailing,
  and consecutive commas before strtok_r tokenization).
- Per-packet match via FLOW_PKT_ESTABLISHED flag (setup vs established)
  and direct TcpSession::state membership check (closing phase).
- SigMatch registration with SIGMATCH_SUPPORT_FIREWALL so the keyword
  works in firewall-mode rules.
- 11 unit tests covering valid parsing, invalid inputs, duplicate
  rejection, match behavior, midstream pickup, and exhaustive
  state x flag x phase_flags verification.
- Engine-analysis JSON output integration.

+++ src/detect.h
@@ -390,7 +390,7 @@ typedef struct TransformData_ {
 
 typedef struct DetectEngineTransforms {
     TransformData transforms[DETECT_TRANSFORMS_MAX];
-    uint8_t cnt;
+    int cnt;
 } DetectEngineTransforms;
 
 /** callback for getting the buffer we need to prefilter/inspect */
